fix: broken embeds
